@@ -1,4 +1,4 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -37,6 +37,9 @@ let package = Package(
                 .product(name: "Algorithms", package: "swift-algorithms"),
                 .product(name: "Atomics", package: "swift-atomics"),
                 .product(name: "NeedleTailLogger", package: "needletail-logger")
+            ],
+            swiftSettings: [
+                .unsafeFlags(["-warn-concurrency"], .when(configuration: .debug))
             ]
         ),
         .target(name: "NeedleTailQueue"),
@@ -48,6 +51,9 @@ let package = Package(
                 .product(name: "NIOConcurrencyHelpers", package: "swift-nio"),
                 .product(name: "BSON", package: "BSON"),
                 .product(name: "NeedleTailLogger", package: "needletail-logger")
+            ],
+            swiftSettings: [
+                .unsafeFlags(["-warn-concurrency"], .when(configuration: .debug))
             ]
         ),
         .testTarget(
